Adds view for all timeline items.
Adds an 'items' view to the timelines controller, with a neatlinetime-json
context. This now means the JSON for a timeline's items is
/neatline-time/timelines/items/1?output=neatlinetime-json.

Removes neatlinetime-json context from ItemsController, since it is not
currently possible to retrieve all the items at once at items/browse. This
will likely make issue #2 a wontfix.