Adds configuration options.
Adds configuration options that lets users choose which fields to use for item
title, date, and description.

Adds helper functions for dealing with configuration options.

Adds tests for checking the setting of these options.

Updates the translation template to account for new strings in the
configuration form.

Adds an upgrade check to add default options.